+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application.Tests/AggregateRoots/GetAggregateRootCompositeManyBByIdQueryHandlerTests.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using AutoFixture;
 using AutoMapper;
 using CleanArchitecture.TestApplication.Application.AggregateRoots.GetAggregateRootCompositeManyBById;
+using CleanArchitecture.TestApplication.Application.Common.Exceptions;
 using CleanArchitecture.TestApplication.Application.Tests.CRUD.AggregateRoots;
 using CleanArchitecture.TestApplication.Domain.Common;
 using CleanArchitecture.TestApplication.Domain.Entities;
@@ -71,7 +72,7 @@ namespace CleanArchitecture.TestApplication.Application.Tests.AggregateRoots
         }
 
         [Fact]
-        public async Task Handle_WithInvalidIdQuery_ReturnsEmptyResult()
+        public async Task Handle_WithInvalidIdQuery_ThrowsNotFoundException()
         {
             // Arrange
             var fixture = new Fixture();
@@ -86,10 +87,10 @@ namespace CleanArchitecture.TestApplication.Application.Tests.AggregateRoots
             var sut = new GetAggregateRootCompositeManyBByIdQueryHandler(repository, _mapper);
 
             // Act
-            var result = await sut.Handle(testQuery, CancellationToken.None);
+            var act = async () => await sut.Handle(testQuery, CancellationToken.None);
 
             // Assert
-            result.Should().Be(null);
+            await act.Should().ThrowAsync<NotFoundException>();
         }
     }
 }