+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application.Tests/AggregateTestNoIdReturns/GetAggregateTestNoIdReturnByIdQueryHandlerTests.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using AutoFixture;
 using AutoMapper;
 using CleanArchitecture.TestApplication.Application.AggregateTestNoIdReturns.GetAggregateTestNoIdReturnById;
+using CleanArchitecture.TestApplication.Application.Common.Exceptions;
 using CleanArchitecture.TestApplication.Application.Tests.CRUD.AggregateTestNoIdReturns;
 using CleanArchitecture.TestApplication.Domain.Common;
 using CleanArchitecture.TestApplication.Domain.Entities;
@@ -67,7 +68,7 @@ namespace CleanArchitecture.TestApplication.Application.Tests.AggregateTestNoIdR
         }
 
         [Fact]
-        public async Task Handle_WithInvalidIdQuery_ReturnsEmptyResult()
+        public async Task Handle_WithInvalidIdQuery_ThrowsNotFoundException()
         {
             // Arrange
             var fixture = new Fixture();
@@ -79,10 +80,10 @@ namespace CleanArchitecture.TestApplication.Application.Tests.AggregateTestNoIdR
             var sut = new GetAggregateTestNoIdReturnByIdQueryHandler(repository, _mapper);
 
             // Act
-            var result = await sut.Handle(query, CancellationToken.None);
+            var act = async () => await sut.Handle(query, CancellationToken.None);
 
             // Assert
-            result.Should().Be(null);
+            await act.Should().ThrowAsync<NotFoundException>();
         }
     }
 }