+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/AggregateRootLongs/GetAggregateRootLongById/GetAggregateRootLongByIdQueryHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using CleanArchitecture.TestApplication.Application.Common.Exceptions;
 using CleanArchitecture.TestApplication.Domain.Repositories;
 using CleanArchitecture.TestApplication.Domain.Repositories.CRUD;
 using Intent.RoslynWeaver.Attributes;
@@ -31,6 +32,11 @@ namespace CleanArchitecture.TestApplication.Application.AggregateRootLongs.GetAg
             CancellationToken cancellationToken)
         {
             var aggregateRootLong = await _aggregateRootLongRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (aggregateRootLong is null)
+            {
+                throw new NotFoundException($"Could not find AggregateRootLong {request.Id}");
+            }
             return aggregateRootLong.MapToAggregateRootLongDto(_mapper);
         }
     }