+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/AggregateTestNoIdReturns/GetAggregateTestNoIdReturnById/GetAggregateTestNoIdReturnByIdQueryHandler.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using CleanArchitecture.TestApplication.Application.Common.Exceptions;
 using CleanArchitecture.TestApplication.Domain.Repositories;
 using CleanArchitecture.TestApplication.Domain.Repositories.CRUD;
 using Intent.RoslynWeaver.Attributes;
@@ -31,6 +32,11 @@ namespace CleanArchitecture.TestApplication.Application.AggregateTestNoIdReturns
             CancellationToken cancellationToken)
         {
             var aggregateTestNoIdReturn = await _aggregateTestNoIdReturnRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (aggregateTestNoIdReturn is null)
+            {
+                throw new NotFoundException($"Could not find AggregateTestNoIdReturn {request.Id}");
+            }
             return aggregateTestNoIdReturn.MapToAggregateTestNoIdReturnDto(_mapper);
         }
     }