+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/ImplicitKeyAggrRoots/GetImplicitKeyAggrRootImplicitKeyNestedCompositionById/GetImplicitKeyAggrRootImplicitKeyNestedCompositionByIdQueryHandler.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using CleanArchitecture.TestApplication.Application.Common.Exceptions;
 using CleanArchitecture.TestApplication.Domain.Entities;
 using CleanArchitecture.TestApplication.Domain.Entities.CRUD;
 using CleanArchitecture.TestApplication.Domain.Repositories;
@@ -40,7 +41,12 @@ namespace CleanArchitecture.TestApplication.Application.ImplicitKeyAggrRoots.Get
             }
 
             var element = aggregateRoot.ImplicitKeyNestedCompositions.FirstOrDefault(p => p.Id == request.Id);
-            return element == null ? null : element.MapToImplicitKeyAggrRootImplicitKeyNestedCompositionDto(_mapper);
+
+            if (element is null)
+            {
+                throw new NotFoundException($"Could not find ImplicitKeyNestedComposition {request.Id}");
+            }
+            return element.MapToImplicitKeyAggrRootImplicitKeyNestedCompositionDto(_mapper);
         }
     }
 }